% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_LESLIE.R
\name{app_LESLIE}
\alias{app_LESLIE}
\title{Run LESLIE Shiny App}
\usage{
app_LESLIE(...)
}
\arguments{
\item{...}{arguments passed to \link[shiny:runApp]{shiny::runApp}}
}
\description{
Runs LESLIE Shiny App
}
\examples{
\dontrun{
app_LESLIE()
}

}
\author{
Sebastian Kreutzer, Ruprecht-Karl-University of Heidelberg (Germany)
}
