% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LESLIE-package.R
\docType{package}
\name{LESLIE-package}
\alias{LESLIE-package}
\alias{LESLIE}
\title{LoESs coLorimetry sIgnal Enhancement\cr
\if{html}{
\figure{LESLIE_logo.png}{options: width="75" alt="r-luminescence.org"}
}}
\description{
Enhance and visualise stratigraphic colour profiles and
decipher hidden features in sediment records.
}
\keyword{package}
