% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LESLIE-package.R
\name{LESLIE_profile}
\alias{LESLIE_profile}
\title{LESLIE example data}
\format{
A four column \link{data.frame}:
}
\description{
LESLIE package example data generated randomly
}
\examples{

data(LESLIE_profile, envir = environment())


}
\keyword{datasets}
