% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_colProfile.R
\name{plot_colProfile}
\alias{plot_colProfile}
\title{Plot the enhanced colour profile}
\usage{
plot_colProfile(
  data,
  depth_top = NULL,
  depth_bottom = NULL,
  cycles = 5L,
  plot = TRUE,
  ...
)
}
\arguments{
\item{data}{\link{data.frame} \link{matrix} (\strong{required}): input data frame with four columns, and each
row defining depth (1 column), and Lab (3 columns) colours. More columns are silently removed.}

\item{depth_top}{\link{numeric} (\emph{optional}): top depth of each layer}

\item{depth_bottom}{\link{numeric} (\emph{optional}): bottom depth of each layer}

\item{cycles}{\link{integer} (\emph{with default}): number of colour enhancement cycles, it also determines
the number of profiles to be plotted}

\item{plot}{\link{logical} (\emph{with default}): enables/disables plot output. If disables, the function
returns a \link{list} with \link{matrix} objects with the colour enhanced profiles}

\item{...}{additional parameters to control the plot output, supported are \code{ylim},\code{ylab}, \code{main}, \code{border_lwd}}
}
\value{
Returns a plot or if \code{plot = FALSE} a \link{list} with \link{matrix} objects
}
\description{
CIELAB color space

Process colour data to display an enhanced colour profile
}
\examples{
data(LESLIE_profile, envir = environment())
plot_colProfile(LESLIE_profile, cycles = 3)

}
