% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_colProfile.R
\name{plot_colProfile}
\alias{plot_colProfile}
\title{Plot the enhanced colour profile}
\usage{
plot_colProfile(
  data,
  depth_top = NULL,
  depth_bottom = NULL,
  cycles = 5L,
  plot = TRUE,
  ...
)
}
\arguments{
\item{data}{\link{data.frame} \link{matrix} (\strong{required}): input data frame with four columns, and each
row defining depth (1 column), and Lab (3 columns) colours. More columns are silently removed.}

\item{depth_top}{\link{numeric} (\emph{optional}): top depth of each layer}

\item{depth_bottom}{\link{numeric} (\emph{optional}): bottom depth of each layer}

\item{cycles}{\link{numeric} (\emph{with default}): parameter for the number of colour enhancement cycles, it also determines the number of profiles for the plot output plotted.
The highest number always determines the number of colour enhancement cycles,
while the vector sequence sets the number of profiles shown in the plot.
If the original sequence (no enhancement) is not wanted in the plot, this should be indicated by \code{-1}.
For instance: '\code{cycles = c(-1,3)} shows only the 3rd enhancement cycle in the plot.}

\item{plot}{\link{logical} (\emph{with default}): enables/disables plot output. If disables, the function
returns a \link{list} with \link{matrix} objects with the colour enhanced profiles}

\item{...}{additional parameters to control the plot output, supported are \code{ylim},\code{ylab}, \code{main}, \code{border_lwd}}
}
\value{
Returns a plot or if \code{plot = FALSE} a \link{list} with \link{matrix} objects
}
\description{
CIELAB colour space

Process colour data to display an enhanced colour profile
}
\examples{
data(LESLIE_profile, envir = environment())
plot_colProfile(LESLIE_profile, cycles = 3)

}
